List users on / API endpoint
Add a new root route (`/`) to list users and the previous timestamp
when they were fetched.

This endpoint should return at most 2 users which points value is greater than the
current `min_number` stored in the UserPointsServer state.

Every time users are fetched, the `timestamp` stored in the GenServer state
should be updated.

Users are fetched asynchronously to avoid blocking the GenServer loop while
the db query is being executed.

We also took the time to refactor tests around the GenServer by manually
starting it inside each test we avoid race conditions around the global state
that is stored.