Add db constraint to ensure points are within the required range
As, for performance reasons, we decided not to use Ecto changesets and
validations to regenerate the users points, we need to ensure the correct
parameters are being sent to the db `random()` function call in
`UScore.Users.regenerate_points/0`. A db constraint comes in handy in such
cases and will help us track any drifts while also preventing users with points
outside the 0..100 range.